backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.3.0
motor==3.7.1
pymongo==4.14.1
python-multipart==0.0.20
pydantic-settings==2.10.1
pillow==11.3.0
pytesseract==0.3.13
google-generativeai==0.8.5
